Add more withdrawal fees to bittrex
I spent a lot of time wondering why the fees didn't seem to load. Noobishly, I tried running asynchronous code as synchronous code and it took me two hours to realize that the fees only update if the system does another process or sleeps. 

I'm a python guy so this is the first time I've looked into javascript with any scrutiny.