Fix Kuna splitting of quote and base currency
Kuna has both USD and USDT denominated markets, so the logic to split up
their markets was failing for certain currencies.  Also, the quote
currencies listed in the file are out of date compared to the current
state of the exchange.  I avoided using `endsWith` because I wasn't sure
if the transpiler can make that expression work.  Let me know if you
have any questions or comments.