Fix ftx's parseTrade function
Problem: in FTX's parseTrades function market object
conditionally overrode by the map of all markets, but
then it's treated as a just market object. Because of this
this function couldn't map symbol from market to trade
object, and then parseTrades filters out trades without
symbol and as a result ccxt returns an empty list of trades.

Solution: Fixed by setting a market variable as an element
of markets_by_id with id of marketId.