Fix setMarkets to set this.markets to the ones gotten from exchange
Problem: Some exchanges have derivatives that could be expired
in a weeks. So after they are expired, the exchange doesn't return this
markets and can't anymore work with them. But in setMarkets function
we are saving old markets and overwriting them with a new ones, so the
expired markets will be presented in ccxt's fetchMarkets result.

Solution: Removed deepExtend function for setting the markets.