Report huobi fetchOrderBook error as BadSymbol
This error occurs when attempting to fetch the order book for a market
on which trading has been suspended. The website returns the following
error, in this case:

> 'Trading is currently suspended. The resumption time will be notified
> via announcement'

If markets are reloaded, the offending market will be marked inactive.
It seems that the issue here is that the market is unavailable for the
provided symbol, and so the error should be `BadSymbol`.