transpile.js add support for python decimal.Decimal
Add support for python's built-in type decimal.Decimal to the transpiling process. Where a number is required in JavaScript, a numbers.Real or a decimal.Decimal is permitted in Python. Unfortunately use of decimal.Decimal is complicated by needing to import it only when it is used, so tools are added to import the library only when necessary.