Simplified the way we deal with symbols
Gooplex uses a distinct format for symbols, which is not the same
everyone else uses. To avoid confusion, we now have two functions to
convert the Gooplex symbol to the common one on the incoming information
and another that does the other way around, so the API can actually find
the symbols.