Update ftx-fetch-all-my-trades.py
just found that when multiple trades happened in the exact same time, it might get skipped. take the following trades as example:
            2601438152  2021-06-13T17:43:46.315639+00:00
            2601438154  2021-06-13T17:43:46.315639+00:00
            2601438156  2021-06-13T17:43:46.315639+00:00
            2601438158  2021-06-13T17:43:46.315639+00:00

if first_trade == 2601438158, then the others 3 trades will be skipped while querying for the next page since the end_time would be "2021-06-13T17:43:46.315639".
since there is no way to query page by trade id, I used a quick and dirty fix to add 1000 milliseconds to make sure the results overlap, and remove the duplicate records in the end.