about "active" status for currency
In my experience, the "status" of currency is not a unified stance across exchanges, sometimes coin might have 'active' status for deposit, but not withdrawal (happens every day on all exchanges, when they list a new token, they allow withdrawals only after specific timespan, typically a day or more, to keep liquidity for a while), or sometimes both (withdrawal & deposit can be available, like it is for GYEN token on binance) but trading is not available.  Thus the 'status' of currency is a combination of different factors, depending exchange API (some exchanges provide separate fields for deposit/withdrawal/trading statuses, and some exchanges just provide only 1 property named 'status/etc' which you cant figure out what it means exactly - trading or funding or what delisted state.

Thus I think, we should be less specific with that property, and just use "or" relativity among those factors. When that property is to 'false' then user should take caution and make further research whether that currency is just disabled for trading, or deposit or withdrawal disabled only.