Fix rounding issue in Luno
The extra float casting is causing the work of e.g., priceToPrecision to be undone, for prices that are very small, such as 0.000016. Casting this to float results in this being sent to Luno, in PHP at least, as 1.6E-5 (instead of as a decimal string), which Luno can't handle.