Fix a problem caused by conflicts between some bitflyer ids
Bitflyer has a different type of conflict in markets, because
some of their ids (ETH/BTC and BTC/JPY) are duplicated in US, EU and JP.
This problem appears when calling methods that use safeMarket internally
(e.g. fetchOrders, fetchOrder).
Fixing it in a way that follows the solution in bithumb solves this problem.