feat(bitfinex2): set the remaining margin methods to false
Set the remaining spot margin related methods that don't have a corresponding endpoint to false.

Someone with access to a bitfinex account will need to make an spot margin order using createOrder and check that it is actually a spot margin order on the website.

As far as I can see it looks like spot orders will have the EXCHANGE prefix in the type when creating an order like `EXCHANGE LIMIT` and spot margin orders don't have a prefix and just use `LIMIT`