test.market.ts allow linear and inverse to be undefined for options (#25927)
Changes to allow linear and inverse to be empty for testing option markets.

For option markets on bybit in some methods when selecting the category param it was set to linear before option in the conditional, to help ensure this doesn't happen the linear and inverse fields in fetchMarkets were set to undefined, but that caused a testing error.