feat(bitget): fetchMarkets UTA
Added UTA support as an optionAndParam, I'm not sure if we want to handle it more like bybit with the isUnifiedEnabled helper method.
Non UTA public endpoints are still available with the UTA api keys.

I'm not sure if we also want to fetch the margin pairs because it overlaps with the spot pairs and they have different limits and precisions.